pyproject.toml
[tool.poetry]
name = "pyramid_grpc"
version = "0.0.1"
description = "Grpc with pyramid"
authors = ["Tomas Correa <ftomas.correa@gmail.com>"]
repository = "https://github.com/tomas_correa/pyramid-grpc"
documentation = "https://tomas_correa.github.io/pyramid-grpc/"
readme = "README.md"
packages = [
  {include = "pyramid_grpc"}
]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
grpc-interceptor = "*"
pyramid = "*"
grpcio = "*"
click = "*"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
deptry = "^0.6.4"
mypy = "^0.981"
pre-commit = "^2.20.0"
tox = "^3.25.1"
pytest-grpc = "^0.8.0"
pyramid-tm = "^2.5"
zope-sqlalchemy = "^3.0"
pyramid-jwt = "^1.6.1"
python-jose = "^3.3.0"
cryptography = "^41.0.3"
grpcio-tools = "^1.57.0"



[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py311']
preview = true

[tool.mypy]
files = ["pyramid_grpc"]
ignore_missing_imports = "True"
disallow_untyped_defs = "False"
disallow_any_unimported = "False"
no_implicit_optional = "True"
check_untyped_defs = "True"
warn_return_any = "True"
warn_unused_ignores = "True"
show_error_codes = "True"

[tool.ruff]
target-version = "py37"
line-length = 120
fix = true
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["pyramid_grpc"]

[tool.poetry.scripts]
grpc-server = 'grpc.grpc.server:run'

[tool.deptry]
extend_exclude = [
    ".venv",
    ".tox"
]

